Phase 2 Session 10/requirements.txt
appnope==0.1.0
astroid==2.3.3
atari-py==0.2.6
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
box2d-py==2.3.8
cloudpickle==1.3.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
gym==0.17.1
importlib-metadata==1.6.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
json5==0.9.4
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.0.1
jupyterlab-server==1.0.7
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.5
notebook==6.0.3
numpy==1.18.2
opencv-python==4.2.0.34
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pybullet==2.7.1
pyglet==1.5.0
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytorch-model-summary==0.1.1
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
terminado==0.8.3
testpath==0.4.4
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
